requirements/base.txt
packaging==16.8
six==1.10.0
appdirs==1.4.3
pylint==1.6.5
pep8==1.7.0
coverage==4.3.4
psycopg2==2.7
djangorestframework==3.7.0
Markdown==2.6.9
django-rest-swagger==2.1.1
django-cors-headers==1.1.0
django-filter==1.0.4
django==1.11.4
pillow==4.2.1
boto==2.43.0
django-ses==0.8.0
pika==0.10.0
python-docx==0.8.6
cloudconvert==1.0.0
django-s3-storage==0.9.10
djangorestframework-filters==0.10.1
djangorestframework-bulk==0.2.1
celery==4.0.1
suds-py3==1.3.2.0
xmltodict==0.10.2
djangorestframework-xml==1.3.0
rollbar==0.13.10
boto3==1.4.4
inflection==0.3.1
python-stdnum==1.5
Faker==0.7.17
zxcvbn==1.0
zxcvbn-python==4.4.16
slackclient==1.0.0
django-cors-middleware==1.3.1
channels_api